requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
pytesseract==0.3.10
opencv-python-headless==4.5.5.64
Pillow==9.5.0
whitenoise==6.4.0
